MonsterObserver

+++ core/src/com/ageofisolation/playScreen/AbstractStatusEffectRenderer.java
@@ -4,8 +4,8 @@ import com.ageofisolation.graphics.GraphicsSingleton;
 import com.ageofisolation.playScreen.Entity.StatusEffectRenderInfo;
 import com.ageofisolation.playScreen.Entity.StatusEffectRenderer;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.badlogic.gdx.math.Rectangle;
 
-import java.awt.*;
 
 public abstract class AbstractStatusEffectRenderer implements StatusEffectRenderer {
     protected int statusEffectCounter;
@@ -18,7 +18,7 @@ public abstract class AbstractStatusEffectRenderer implements StatusEffectRender
         Rectangle rectangle = getRenderRectangle();
         SpriteBatch spriteBatch = GraphicsSingleton.getInstance().getSpriteBatch();
         spriteBatch.draw(GraphicsSingleton.getInstance().getTexture(info.statusEffectTexturePath()),
-                (float) rectangle.getX(), (float) rectangle.getY(),(float) rectangle.getWidth(),(float) rectangle.getHeight());
+                rectangle.getX(), rectangle.getY(), rectangle.getWidth(), rectangle.getHeight());
         //TODO: render stacks amount
     }
